added BigDecimals for integrals

+++ src/edu/duke/cs/osprey/partitionfunctionbounds/continuous/SCMF.java
@@ -215,7 +215,7 @@ public class SCMF {
 				nDomain.domainUB,
 				(nPoint)->(ed.eFunc.applyAsDouble(CMRFEdgeDomain.concatArrays(point, nPoint))));
 
-		return func.computeExpectation();
+		return func.computeExpectation().doubleValue();
 	}
 
 	/**
